requirements.txt
wheel
Django==3.1.3
django-mysql
django-crontab
mysqlclient
pyyaml~=5.3.1
requests
paramiko
djangorestframework==3.12.2
django-cors-headers
jira
pika==1.1.0